+++ python/metadata.cc
@@ -10,6 +10,7 @@
 #include "arki/types/source.h"
 #include "arki/structured/keys.h"
 #include "arki/structured/json.h"
+#include "arki/structured/memory.h"
 #include "arki/formatter.h"
 #include "utils/core.h"
 #include "utils/methods.h"
@@ -454,7 +455,6 @@ struct read_bundle : public ClassMethKwargs<read_bundle>
     }
 };
 
-
 struct write_bundle : public ClassMethKwargs<write_bundle>
 {
     constexpr static const char* name = "write_bundle";
@@ -487,6 +487,143 @@ struct write_bundle : public ClassMethKwargs<write_bundle>
     }
 };
 
+struct read_yaml : public ClassMethKwargs<read_yaml>
+{
+    constexpr static const char* name = "read_yaml";
+    constexpr static const char* signature = "src: Union[str, StringIO, bytes, ByteIO]";
+    constexpr static const char* returns = "arkimet.Metadata";
+    constexpr static const char* summary = "Read a Metadata from a YAML file";
+
+    static PyObject* run(PyTypeObject* cls, PyObject* args, PyObject* kw)
+    {
+        static const char* kwlist[] = { "src", nullptr };
+        PyObject* py_src = nullptr;
+        if (!PyArg_ParseTupleAndKeywords(args, kw, "O", const_cast<char**>(kwlist), &py_src))
+            return nullptr;
+
+        try {
+            std::unique_ptr<Metadata> res(new Metadata);
+            if (PyBytes_Check(py_src))
+            {
+                char* buffer;
+                Py_ssize_t length;
+                if (PyBytes_AsStringAndSize(py_src, &buffer, &length) == -1)
+                    throw PythonException();
+                ReleaseGIL gil;
+                auto reader = arki::core::LineReader::from_chars(buffer, length);
+                res->readYaml(*reader, "bytes buffer");
+            } else if (PyUnicode_Check(py_src)) {
+                Py_ssize_t length;
+                const char* buffer = throw_ifnull(PyUnicode_AsUTF8AndSize(py_src, &length));
+                ReleaseGIL gil;
+                auto reader = arki::core::LineReader::from_chars(buffer, length);
+                res->readYaml(*reader, "str buffer");
+            } else if (PyObject_HasAttrString(py_src, "encoding")) {
+                TextInputFile input(py_src);
+                ReleaseGIL gil;
+
+                std::unique_ptr<arki::core::LineReader> reader;
+                std::string input_name;
+                if (input.fd)
+                {
+                    input_name = input.fd->name();
+                    reader = arki::core::LineReader::from_fd(*input.fd);
+                }
+                else
+                {
+                    input_name = input.abstract->name();
+                    reader = arki::core::LineReader::from_abstract(*input.abstract);
+                }
+
+                res->readYaml(*reader, input_name);
+            } else {
+                BinaryInputFile input(py_src);
+                ReleaseGIL gil;
+
+                std::unique_ptr<arki::core::LineReader> reader;
+                std::string input_name;
+                if (input.fd)
+                {
+                    input_name = input.fd->name();
+                    reader = arki::core::LineReader::from_fd(*input.fd);
+                }
+                else
+                {
+                    input_name = input.abstract->name();
+                    reader = arki::core::LineReader::from_abstract(*input.abstract);
+                }
+                res->readYaml(*reader, input_name);
+            }
+            return (PyObject*)metadata_create(std::move(res));
+        } ARKI_CATCH_RETURN_PYO
+    }
+};
+
+struct read_json : public ClassMethKwargs<read_json>
+{
+    constexpr static const char* name = "read_json";
+    constexpr static const char* signature = "src: Union[str, StringIO, bytes, ByteIO]";
+    constexpr static const char* returns = "arkimet.Metadata";
+    constexpr static const char* summary = "Read a Metadata from a JSON file";
+
+    static PyObject* run(PyTypeObject* cls, PyObject* args, PyObject* kw)
+    {
+        static const char* kwlist[] = { "src", nullptr };
+        PyObject* py_src = nullptr;
+        if (!PyArg_ParseTupleAndKeywords(args, kw, "O", const_cast<char**>(kwlist), &py_src))
+            return nullptr;
+
+        try {
+            arki::structured::Memory parsed;
+            if (PyBytes_Check(py_src))
+            {
+                char* buffer;
+                Py_ssize_t length;
+                if (PyBytes_AsStringAndSize(py_src, &buffer, &length) == -1)
+                    throw PythonException();
+                auto input = arki::core::BufferedReader::from_chars(buffer, length);
+                ReleaseGIL gil;
+                arki::structured::JSON::parse(*input, parsed);
+            } else if (PyUnicode_Check(py_src)) {
+                Py_ssize_t length;
+                const char* buffer = throw_ifnull(PyUnicode_AsUTF8AndSize(py_src, &length));
+                auto input = arki::core::BufferedReader::from_chars(buffer, length);
+                ReleaseGIL gil;
+                arki::structured::JSON::parse(*input, parsed);
+            } else if (PyObject_HasAttrString(py_src, "encoding")) {
+                TextInputFile input(py_src);
+
+                std::unique_ptr<arki::core::BufferedReader> reader;
+                if (input.fd)
+                    reader = arki::core::BufferedReader::from_fd(*input.fd);
+                else
+                    reader = arki::core::BufferedReader::from_abstract(*input.abstract);
+
+                ReleaseGIL gil;
+                arki::structured::JSON::parse(*reader, parsed);
+            } else {
+                BinaryInputFile input(py_src);
+
+                std::unique_ptr<arki::core::BufferedReader> reader;
+                if (input.fd)
+                    reader = arki::core::BufferedReader::from_fd(*input.fd);
+                else
+                    reader = arki::core::BufferedReader::from_abstract(*input.abstract);
+
+                ReleaseGIL gil;
+                arki::structured::JSON::parse(*reader, parsed);
+            }
+
+            ReleaseGIL gil;
+            std::unique_ptr<Metadata> res(new Metadata);
+            res->read(arki::structured::keys_json, parsed.root());
+
+            gil.lock();
+            return (PyObject*)metadata_create(std::move(res));
+        } ARKI_CATCH_RETURN_PYO
+    }
+};
+
 
 struct MetadataDef : public Type<MetadataDef, arkipy_Metadata>
 {
@@ -514,7 +651,7 @@ For example::
     GetSetters<data, data_size> getsetters;
     Methods<
         has_source, write, make_absolute, make_inline, make_url, to_string,
-        to_python, get_notes, del_notes, read_bundle, write_bundle> methods;
+        to_python, get_notes, del_notes, read_bundle, write_bundle, read_yaml, read_json> methods;
 
     static void _dealloc(Impl* self)
     {